commercial/requirements.txt
# 商业化模块依赖

# 核心框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2

# 支付
alipay-sdk-python>=3.7.0
wechatpy>=1.8.18
cryptography>=41.0.0

# 工具
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
redis>=5.0.0
python-dotenv>=1.0.0
